Create @amplication/git-service web app #2462

+++ packages/amplication-server/src/core/git/git.wrapper.service.ts
@@ -15,25 +15,26 @@ import { ConnectGitRepositoryInput } from './dto/inputs/ConnectGitRepositoryInpu
 import { CreateGitRepositoryInput } from './dto/inputs/CreateGitRepositoryInput';
 import { RemoteGitRepositoriesWhereUniqueInput } from './dto/inputs/RemoteGitRepositoriesWhereUniqueInput';
 import { RemoteGitRepository } from './dto/objects/RemoteGitRepository';
-import { GitServiceFactory } from './utils/GitServiceFactory/GitServiceFactory';
-import { EnumGitOrganizationType } from './dto/enums/EnumGitOrganizationType';
+import { GitService } from '@amplication/git-service/src/services/git.service';
+import { EnumGitOrganizationType } from '@amplication/git-service/src/Dto/enums/EnumGitOrganizationType';
 
 @Injectable()
-export class GitService {
+export class GitWrapperService {
   constructor(
-    private readonly gitServiceFactory: GitServiceFactory,
-    private readonly prisma: PrismaService
+    private readonly prisma: PrismaService,
+    private readonly gitService: GitService
   ) {}
 
   async getReposOfOrganization(
     args: RemoteGitRepositoriesWhereUniqueInput
   ): Promise<RemoteGitRepository[]> {
-    const { gitProvider, gitOrganizationId } = args;
-    const gitService = this.gitServiceFactory.getService(gitProvider);
     const installationId = await this.getInstallationIdByGitOrganizationId(
-      gitOrganizationId
+      args.gitOrganizationId
     );
-    return await gitService.getOrganizationRepos(installationId);
+    return await this.gitService.getReposOfOrganization(
+      args.gitProvider,
+      installationId
+    );
   }
 
   async createGitRepository(args: CreateGitRepositoryInput): Promise<App> {
@@ -43,19 +44,13 @@ export class GitService {
       }
     });
 
-    const provider = this.gitServiceFactory.getService(args.gitProvider); //TODO: organization.provider
-    const repository = await (organization.type ===
-    EnumGitOrganizationType.Organization
-      ? provider.createOrganizationRepository(
-          organization.installationId,
-          organization.name,
-          args.name
-        )
-      : provider.createUserRepository(
-          organization.installationId,
-          organization.name,
-          args.name
-        ));
+    const repository = await this.gitService.createGitRepository(
+      args.name,
+      args.gitProvider,
+      EnumGitOrganizationType[organization.type],
+      organization.name,
+      organization.installationId
+    );
 
     if (!repository) {
       throw new AmplicationError(
@@ -117,10 +112,10 @@ export class GitService {
     args: CreateGitOrganizationArgs
   ): Promise<GitOrganization> {
     const { gitProvider, installationId } = args.data;
-    const service = this.gitServiceFactory.getService(gitProvider);
 
-    const gitRemoteOrganization = await service.getGitRemoteOrganization(
-      installationId
+    const gitRemoteOrganization = await this.gitService.getGitRemoteOrganization(
+      installationId,
+      gitProvider
     );
 
     const gitOrganization = await this.prisma.gitOrganization.findFirst({
@@ -173,25 +168,34 @@ export class GitService {
     args: GetGitInstallationUrlArgs
   ): Promise<string> {
     const { gitProvider, workspaceId } = args.data;
-    const service = this.gitServiceFactory.getService(gitProvider);
-    return await service.getGitInstallationUrl(workspaceId);
+    return await this.gitService.getGitInstallationUrl(
+      gitProvider,
+      workspaceId
+    );
   }
 
   async deleteGitOrganization(
     args: DeleteGitOrganizationArgs
   ): Promise<boolean> {
     const { gitProvider, gitOrganizationId } = args;
-    const service = this.gitServiceFactory.getService(gitProvider);
     const installationId = await this.getInstallationIdByGitOrganizationId(
       gitOrganizationId
     );
-    await service.deleteGitOrganization(installationId);
-    await this.prisma.gitOrganization.delete({
-      where: {
-        id: gitOrganizationId
+    if (installationId) {
+      const isDelete = await this.gitService.deleteGitOrganization(
+        gitProvider,
+        installationId
+      );
+      if (isDelete) {
+        await this.prisma.gitOrganization.delete({
+          where: {
+            id: gitOrganizationId
+          }
+        });
+        return true;
       }
-    });
-    return true;
+    }
+    return false;
   }
 
   private async getInstallationIdByGitOrganizationId(
